Move GenericAgentProxy to proxies
