+++ emote/algorithms/sac.py
@@ -9,11 +9,10 @@ import torch
 
 from torch import nn, optim
 
-from emote.algorithms.generic_agent_proxy import GenericAgentProxy
 from emote.callback import Callback
 from emote.callbacks.loss import LossCallback
 from emote.mixins.logging import LoggingMixin
-from emote.proxies import AgentProxy
+from emote.proxies import AgentProxy, GenericAgentProxy
 from emote.utils.deprecated import deprecated
 from emote.utils.gamma_matrix import discount, make_gamma_matrix, split_rollouts
 from emote.utils.spaces import MDPSpace
